+++ include/solvers/physical_properties_manager.h
@@ -104,7 +104,8 @@ public:
     for (unsigned int f = 0; f < number_of_fluids; ++f)
       specific_heat[f]->provide_simulation_control(simulation_control);
     for (unsigned int s = 0; s < number_of_solids; ++s)
-      specific_heat[s]->provide_simulation_control(simulation_control);
+      specific_heat[s + number_of_fluids]->provide_simulation_control(
+        simulation_control);
   }
 
   inline unsigned int