+++ source/fem-dem/gls_sharp_navier_stokes.cc
@@ -28,7 +28,7 @@ GLSSharpNavierStokesSolver<dim>::GLSSharpNavierStokesSolver(
   : GLSNavierStokesSolver<dim>(p_nsparam.cfd_parameters)
   , cfd_dem_parameters(p_nsparam)
   , all_spheres(true)
-
+  , combined_shapes()
 {}
 
 template <int dim>
@@ -809,6 +809,16 @@ GLSSharpNavierStokesSolver<dim>::define_particles()
     particles);
 
   check_whether_all_particles_are_sphere();
+
+  std::vector<std::shared_ptr<Shape<dim>>> all_shapes;
+  for (const IBParticle<dim> &particle : particles)
+    {
+      all_shapes.push_back(particle.shape);
+    }
+  combined_shapes =
+    std::make_shared<CompositeShape<dim>>(all_shapes, Point<dim>(), Point<3>());
+  this->multiphysics->set_immersed_solid_signed_distance_function(
+    combined_shapes);
 }
 
 
@@ -1662,19 +1672,6 @@ template <int dim>
 void
 GLSSharpNavierStokesSolver<dim>::output_field_hook(DataOut<dim> &data_out)
 {
-  std::vector<std::shared_ptr<Shape<dim>>> all_shapes;
-  for (const IBParticle<dim> &particle : particles)
-    {
-      all_shapes.push_back(particle.shape);
-    }
-  std::shared_ptr<Shape<dim>> combined_shapes;
-  if (particles.size() == 1)
-    combined_shapes = particles[0].shape;
-  else
-    combined_shapes = std::make_shared<CompositeShape<dim>>(all_shapes,
-                                                            Point<dim>(),
-                                                            Point<3>());
-
   levelset_postprocessor =
     std::make_shared<LevelsetPostprocessor<dim>>(combined_shapes);
   data_out.add_data_vector(this->present_solution, *levelset_postprocessor);
@@ -4942,7 +4939,7 @@ GLSSharpNavierStokesSolver<dim>::solve()
           this->iterate();
         }
 
-      this->postprocess_fd(false);
+      this->postprocess(false);
 
       if (this->simulation_parameters.particlesParameters->calculate_force_ib)
         force_on_ib();